refactor(#258): unify frontend problem-related types without a god type
- Add frontend/src/types/problem.ts with ProblemDetail and ProblemListItem
- Import shared types into ProblemDetailPage, CoachingPage, and ProblemsPage
- Move PracticeHistoryResponse to types/practice.ts
- Remove duplicated inline interfaces from target pages
